Refactored all models according to discussed implementation plan.
