+++ README.md
@@ -1,9 +1,11 @@
 # LITTIL initiative
 
 The LITTIL project is an initiative to develop and maintain an Open-Source platform to connect schools with volunteer
-teachers for programming workshops. It was born to facilitate the central aim of the Devoxx4Kids foundation, which is to
+guestTeachers for programming workshops. It was born to facilitate the central aim of the Devoxx4Kids foundation, which
+is to
 acquaint children with programming through fun workshops at their school.
-The platform provides a web-based portal for schools and volunteer teachers to register themselves and create a public
+The platform provides a web-based portal for schools and volunteer guestTeachers to register themselves and create a
+public
 profile. The platform then facilitates both parties to find possible matches.
 
 ## LITTIL backend API
