+++ src/main/java/org/littil/api/auth/Role.java
@@ -1,0 +1,4 @@
+package org.littil.api.auth;
+
+public record Role(String id, String role) {
+}
