--- src/main/java/org/littil/api/auth/RoleEntity.java
@@ -1,12 +1,0 @@
-package org.littil.api.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class RoleEntity {
-
-	private String id;
-	private String role;
-}
