--- src/main/java/org/littil/api/auth/UserType.java
@@ -1,6 +1,0 @@
-package org.littil.api.auth;
-
-public enum UserType {
-    USER,
-    LITTIL
-}
