+++ src/main/java/org/littil/api/auth/provider/auth0/RoleMapper.java
@@ -1,0 +1,22 @@
+package org.littil.api.auth.provider.auth0;
+
+import org.littil.api.auth.Role;
+
+import javax.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+//TODO implement MapStruct
+public class RoleMapper {
+
+    com.auth0.json.mgmt.Role toDomain(Role role) {
+        com.auth0.json.mgmt.Role mgmtRole = new com.auth0.json.mgmt.Role();
+        mgmtRole.setName(role.role());
+        return mgmtRole;
+    }
+
+    ;
+
+    Role toEntity(com.auth0.json.mgmt.Role role) {
+        return new Role(role.getId(), role.getName());
+    }
+}
