--- src/main/java/org/littil/api/auth/service/TeacherUser.java
@@ -1,5 +1,0 @@
-package org.littil.api.auth.service;
-
-public class TeacherUser extends User {
-
-}
