--- src/main/java/org/littil/api/auth/service/User.java
@@ -1,10 +1,0 @@
-package org.littil.api.auth.service;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotEmpty;
-
-public abstract class User {
-    @NotEmpty(message = "{User.firstName.required}")
-    @Email
-    private String email;
-}
