+++ src/main/java/org/littil/api/location/repository/LocationEntity.java
@@ -1,8 +1,10 @@
 package org.littil.api.location.repository;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.hibernate.validator.constraints.Length;
 import org.littil.api.auditing.repository.AbstractAuditableEntity;
 
 import javax.persistence.Column;
@@ -22,7 +24,7 @@ import java.util.UUID;
 
 @Data
 @NoArgsConstructor
-@AllArgsConstructor
+@RequiredArgsConstructor
 @Entity(name = "Location")
 @Table(name = "location")
 public class LocationEntity extends AbstractAuditableEntity {
@@ -30,14 +32,23 @@ public class LocationEntity extends AbstractAuditableEntity {
     @Id
     @GeneratedValue
     @Column(name = "location_id", columnDefinition = "BINARY(16)")
+    @NonNull
     private UUID id;
 
+    // Code based on ISO-3166
+    @Column(name = "country_code", columnDefinition = "VARCHAR(2)")
+    @Length(max = 2)
+    private String country = "NL";
+
     @NotEmpty(message = "{Location.address.required}")
+    @NonNull
     @Column(name = "address")
     private String address;
 
     @NotEmpty(message = "{Location.postalCode.required}")
+    @NonNull
     @Column(name = "postal_code")
+    @Length(max = 10)
     private String postalCode;
 
     @Column(name = "latitude")
