+++ src/main/resources/ValidationMessages.properties
@@ -1,10 +1,10 @@
 System.error=An unexpected error has occurred. Please contact support.
-Teacher.firstName.required=The name of the teacher is required.
-Teacher.surname.required=The surname of the teacher is required.
-Teacher.email.invalid=The email address of the teacher is invalid.
-Teacher.email.required=The email address of the teacher is required.
-Teacher.postalCode.required=The postal code of the teacher is required.
-Teacher.locale.required=The locale of the teacher is required.
+GuestTeacher.firstName.required=The name of the guest teacher is required.
+GuestTeacher.surname.required=The surname of the guestt eacher is required.
+GuestTeacher.email.invalid=The email address of the guest teacher is invalid.
+GuestTeacher.email.required=The email address of the guest teacher is required.
+GuestTeacher.postalCode.required=The postal code of the guest teacher is required.
+GuestTeacher.locale.required=The locale of the guest teacher is required.
 School.name.required=The name of the school is required.
 School.address.required=The address of the school is required.
 School.postalCode.required=The postal code of the school is required.
