+++ src/main/resources/db/migration/V2_0_0__Create_School_Entity.sql
@@ -4,9 +4,9 @@ CREATE TABLE school
     school_name         VARCHAR(255) NOT NULL,
     contact_person_name VARCHAR(255) NOT NULL,
     location            BINARY(16)   NOT NULL,
-    created_by          VARCHAR(255),
+    created_by          BINARY(16),
     created_date        DATETIME,
-    last_modified_by    VARCHAR(255),
+    last_modified_by    BINARY(16),
     last_modified_date  DATETIME,
     PRIMARY KEY (school_id),
     CONSTRAINT fk_school_location
